Fuck gulp (#116)
gulp has multiple listed vulnerabilities with npm and gulp 4 is still in "next".  gulp was always overkill for what we were doing anyway.  Switch to a small reinvented wheel to just copy over the deps we need, and wire it to prebuild.  End luser will need to ensure node is installed and npm install has run first