WIP TagHelper injection prevention
The idea is to build helper methods like this for all possible escaping contexts in a taghelper, and then use those helper method in every location where we're manipulating string variables in a taghelper.

It's certainly true that certain instances of interpolation can be staticly known to be safe, but that's a hard and unscalable problem to answer in general; you need to think to every variable, where it came from, what could have touched or edited it, etc.