ParamHelper:  Limit valid parameter names
Let's say for now that, from a syntactic perspective, we are only supporting [a-zA-Z] in parameter names.  Other characters can be added later if needed.  The empty string is not supported either; unless we find a place where the old site uses [module:boop|=zoo] and we want that to parse as "" having value "zoo", there's no need to add this curiosity ourselves.

All of the currently used param names meet these requirements.

Tighten up the ParamHelper to throw an exception when a param name does not meet the requirements.  These aren't coming from the user; they're typically constants in the code, so a bad one indicates a real code problem.  The ParamHelper also no longer accepts `null` for a parameter name, since there's nothing meaningful that the function could do with that value.