Remove some copies when running wiki tag helper
I'm a bit out of water on this one.  There aren't good impementing gudies for TextWriter or for whatever is going on with all this TagHelperOutput stuff.

What we'd really like to do is implement IHtmlContent and then pass that TextWriter directly to Util.RenderHtmlAsync, but the problem there is that IHtmlContent doesn't allow async waits.