Remove the keepIdentity flag from the BulkInserter altogether.
With everything cleaned up we're consistently using the existence of an "Id" column in the list of columns to blit to mean that that column is to be kept.  This is cleaner, as neither keepIdentity: true but no identity column copied or keepIdentity: false but an identity column copied made any sense.  It's also needed for postgres bulk copy, which works differently:  It has no special keepIdentity setting, instead going by whatever columns are offered.