JRSR: enhance error checking.
Parse integers with a 0 NumberStyles, so as to reject leading and
trailing whitespace, sign prefixes, decimal points, exponents, etc.
The tasvideos.org JRSR documentation, as of Revision 7, only permits a
'-' prefix for negative numbers, and not a '+' prefix for positive
numbers, but JPC-RR uses e.g. Long.parseLong from Java, which permits
both.
https://repo.or.cz/jpcrr.git/blob/6ab255ce10b2dd10b0ac3ab2e2be708e0b26eeaa:/org/jpc/emulator/EventRecorder.java#l400
https://docs.oracle.com/javase/7/docs/api/java/lang/Long.html#parseLong(java.lang.String)

Throw an error when a line does not contain the expected number of
tokens, instead of ignoring the line.

Check for arithmetic overflow in event timestamps and the computation of
the frame count.