fix(emails): Users with email not confirmed can't request a resend
If a new user is registering and due to site errors is not able to receive the confirmation email but was still registered in the database, then they are not able to login, even if an admin gives login permissions, since only users with email confirmed can login. Additionally, if the user is registering and then closes the browser, they wouldn't be able to access the EmailConfirmationSent page either. For this reason, an update was made so that if users do not have their email confirmed but they exist in the database, then they will be redirected to the EmailConfirmationSent page, so that they can request a resend, only if the username and email entered match.