use webOptimizer Production settings for all other environments as well
These settings control whether the WebOptimizer package adds a cache-control header, and whether the css files get a version string in the html. The Production settings (unchanged) allow browser caching and add a version string. The Dev settings used to disable browser caching and therefore there was no need for a version string. But that's not true, because browsers cached the files anyway. So developers ended up having to Ctrl+F5 to see changes.
With this change, all environments now work exactly as Production does, allow caching, and use version strings for cache-busting on change.