Mark all links in forum markup, plus sub notes and homepages, as UGC
The way this is implemented is that *only external links* are marked as UGC, not links to pubs/wiki/other pages within the site.
By "all links in forum markup" I mean exactly that--every link which appears in everyone's posts in every thread, and all signatures, and all PMs. That simplicity meant it was a 1-line change, compared to all the work it took to distinguish wiki pages.