improve forgot password email, and the page
We now show the username in the email and page, because it's a useful thing that was completely missing previously. The email also contains the usual text to ignore the email if this is unwanted (someone guessed their email), and we also include a note that shows how long the link is valid. 24 hours is the default.