fix a weird primary constructor warning.
the problem was in the ViewModel that inherits the SubmitPageModelBase. it was using the "fileService" of its own constructor, which causes a field to be created in both the base class and the inheriting class. to avoid this, we can conveniently hide the primary constructor "fileService", and instead use our own manual field "fileService", which is then used in both base class and inheriting class.