playground labels now use title override if they clash with movie labels
also display baseline label like it's displayed for movies: (baseline) instead of "baseline"

disambiguate between similar PG goals with different title overrides

overrides are only used to resolve such clashes, for cases when something is accepted under an existing goal label but as a different branch. for example if the version difference warrants it. so we don't have to determine whether an override is meant to be used under the same goal or a different one: it always implies the branch is different. so we spell out the version like we do for pub tabs.